requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.2.0
cssutils==2.11.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
lxml==5.3.1
more-itertools==10.6.0
passlib==1.7.4
premailer==3.10.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yagmail==0.15.293
